Nuked the project structure
First of all: While assets are public domain, code is MIT-license. I prefer GPL. So - I destroyed the code so I can switch to GPL later on.
Second of all: The code was written for 1.12 Forge. 1.13+ Forge is vastly different from 1.12 and maybe only 15% of the code would be compatible. And I'm aming for 1.13+ Fabric. Fabric is not even remotely close to Forge API-wise. There's no way in hell that more than 5% of code will be compatible. So, I'd just feel more comfortable with a complete re-write.
Third of all: The code is a mixture of Java, Scala and LUA. I want pure Java and MAYBE Python instead of Lua.